refactor: 바, 라인 차트마다 라벨 다르게 수정
