+++ src/components/YearTax/Chart/BarChart.jsx
@@ -9,6 +9,7 @@ import {
   Legend,
 } from "chart.js";
 import { Bar } from "react-chartjs-2";
+import { formatDateTime } from "~/lib/utils/dateFormat";
 
 ChartJS.register(
   CategoryScale,
@@ -20,6 +21,7 @@ ChartJS.register(
 );
 
 export const options = {
+  indexAxis: "y",
   responsive: true,
   plugins: {
     legend: {
@@ -33,5 +35,25 @@ export const options = {
 };
 
 export function BarChart({ data }) {
-  return <Bar options={options} data={data} />;
+  const transformedData = {
+    labels: data.map((item) => formatDateTime(new Date(item.createdDate))),
+    datasets: [
+      {
+        label: "돌려받은 돈",
+        data: data.map((item) => item.돌려받은돈),
+        borderColor: "rgb(255, 99, 132)",
+        backgroundColor: "rgba(255, 99, 132, 0.5)",
+        yAxisID: "y",
+      },
+      {
+        label: "총급여",
+        data: data.map((item) => item.총급여),
+        borderColor: "rgb(53, 162, 235)",
+        backgroundColor: "rgba(53, 162, 235, 0.5)",
+        yAxisID: "y",
+      },
+    ],
+  };
+
+  return <Bar options={options} data={transformedData} />;
 }
